v1.0.2 Fix #1 Added an null atacker check to the player death event to prevent NullRefereceException errors
thx Totoro for the error report :D

Full changelog:
Fix formatting and null check in Player Death method

Improved code readability by adding a space after the `if` keyword
in the `Death` method of `Player.cs`. Added a null check to ensure
`attacker` is assigned `player` if it is `null`. No other logic
was altered.